fix text frame

+++ GPImageEditor/Classes/StickerPage/StickerView.swift
@@ -406,7 +406,7 @@ public class StickerView: UIView {
         }
         textView.text = info.text
         let fontConfig = GPImageEditorConfigs.fontSet[info.fontIndex]
-        textView.font = UIFont(name: fontConfig.font, size: fontConfig.size)
+        textView.font = UIFont(name: fontConfig.font, size: CGFloat(fontConfig.size))
         GPTextEditorView.buildTextView(textView)
     }
